+++ src/testHelpers/mockSources/analytics.ts
@@ -1,5 +1,5 @@
 import { Stream } from "xstream";
-import { AnalyticsSource } from "../drivers/analyticsDriver";
+import { AnalyticsSource } from "../../drivers/analyticsDriver";
 
 export const mockAnalyticsSource = (source: {
   anonymousId$?: Stream<string>;
